refactor: 인터페이스 테스트에서 구현체 테스트로 변경
